+++ cmd/vigilante/cmd/root.go
@@ -14,6 +14,8 @@ func NewRootCmd() *cobra.Command {
 		GetSubmitterCmd(),
 		GetMonitorCmd(),
 		GetBTCStakingTracker(),
+		CommandDumpConfig(),
+		CommandVersion(),
 	)
 
 	return rootCmd